button to leave Game and return to Dashboard #20, #16
