+++ src/components/views/SessionResult.js
@@ -1,31 +1,11 @@
 import TableFinalRanking from "../ui/GameSessionResult/TableFinalRanking";
 import * as React from "react";
 import Button from "../ui/Button";
-import {api_with_token, handleError} from "../../helpers/api";
 import {useHistory} from "react-router-dom";
-import LocalStorageManager from "../../helpers/LocalStorageManager";
+import {LeaveGame} from "../../helpers/Utilities";
 
 const SessionResult = () => {
     const history = useHistory();
-
-    const goToDashboard = async () => {
-        try {
-            const username = localStorage.getItem("username");
-            const gameID = localStorage.getItem("gameID");
-            const requestBody = JSON.stringify({
-                username
-            })
-            await api_with_token().post("/games/" + gameID + "/leave", requestBody);
-            history.push("/dashboard");
-            LocalStorageManager.LeaveGame();
-
-        } catch (error) {
-            console.error(`Error while leaving the GameSession: \n${handleError(error)}`);
-            console.error("Details:", error);
-            alert("Error while leaving the GameSession.");
-
-        }
-    }
 
     return (
         <div className="db primary-container">
@@ -34,7 +14,7 @@ const SessionResult = () => {
                 <TableFinalRanking />
                 <br/>
                 <Button
-                    onClick={() => goToDashboard()}>
+                    onClick={() => LeaveGame(history)}>
                     Return to Dashboard
                 </Button>
             </div>
